pyproject.toml
[tool.poetry]
name = "pystan-environment"
version = "0.1.0"
description = ""
authors = ["sthysel <sthysel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pystan = "^3.7.0"
numpy = "^1.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
